// frontend/src/api/generated/contracts.ts
export interface SourceKey {
  "sourceInstanceId": string;
  "collectionId": string;
  "sourceKind": SourceKeySourceKind;
  "externalId": string;
}

export type SourceKeySourceKind = "homebox-entity" | "network-device" | "network-group" | "network-interface" | "network-segment" | "magicplan-room";

export interface SourceRef {
  "workspaceId": string;
  "homeId": string;
  "key": SourceKey;
}